+++ studio/components/interfaces/Settings/Logs/UpgradePrompt.tsx
